frontend/frontend/backend/requirements.txt
# Python Database Testing Requirements
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
tabulate>=0.9.0
